refactor[#40] : entities/tag 에 존재하던 모든 api 요청 하나의 파일로 모듈화
