+++ blog/src/entities/tag/model/tagQueryKey.ts
@@ -1,0 +1,3 @@
+export const TAG_QUERY_KEY = {
+  default: () => ["tags"] as const
+};
